Fixed the width and alginment computation for MMSafe pointer.
On x86-64, Clang assumes that pointer's width is 64 bits and
alginment is 64 bits. This assumption does not hold for
our new MMSafe pointers.

Currently we added new functions in ASTContext.to and
Basic/TargetInfo.h to check the type of the target pointer
and return correct pointer size and alignment. This may not be
the most clean way to implement this. We need to think about
where is the best location to do the poionter checking and computation.

Currently we only implement for x86-64 architecture.