Refactored the compiler to rename _MMSafe_ptr.
Renamed _MMSafe_ptr to _TS_ptr. "TS" stands for Temporal Safety.
The main reason for this change is for brevity. However, I don't
think this change is satisfying. It might be better if we
just do "_TPtr" becausse it saves one underline; for temporal
memory safe version of _Nt_array_ptr, "_TS_Nt_array_ptr" is
tedious to write. We may need to refactor the compiler with
a better name later.