Return _MM_array_ptr<T> to assign to a _MM_array_ptr<concrete_type>.
Added and changed code to handle calling a function that returns a
_MM_array_ptr<T> with a generic type, and also assigning the
return value to a _MM_array_ptr<T> with a concrete data type.
Naturally there would be a type mismatch between the generic type
{i8*, i64, i8*} and the concrete type {ptr_ty, i64, ptr_ty}.
In EmitCall() in CGCall.cpp, mutate the type of the concrete
_MM_array_ptr to be the generic one as the callee's.
For assignment, in EmitStoreOfScalar() in CGExpr.cpp,
mutate the generic type to be the concrete type.

Also in TargetInfo.cpp, updated code of the classification algorithm
for return values and parameters for X86-64 ABI.
_MM_array_ptr has three fields totalling three eight bytes, and
thus it should be passed by memory intead of by register.

As an example, the following Checked C code

    mm_array_ptr<int> p = mm_array_alloc<int>(sizeof(int) * 10);

would generate LLVM IR

    %p = alloca { i32*, i64, i32* }, align 32
    %tmp = alloca { i32*, i64, i32* }, align 32
    call void @mm_array_alloc({ i8*, i64, i8* }* sret %tmp, i64 40)
    %0 = load { i32*, i64, i32* }, { i8*, i64, i8* }* %tmp, align 32
    store { i32*, i64, i32* } %0, { i32*, i64, i32* }* %p, align 32

Currently it compiles on -O0 but fails on -O1 or higher because of
an assertion failure in the SROA pass. It is caused by the type mismatch
of the load instruction.

The mm_array_alloc() implementation is at:
https://github.com/jzhou76/checkedc-safe-mm-misc/commit/e81bce1ad628aa1a311ac61ee85507b6c47bdc84