Updated Address.h to handle _MM_array_ptr.
Added code in Address.h to set the Pointer of an Address that
represents a _MM_array_ptr to be the inner pointer of its struct
implementation. Without this change, pointer dereference would fail
because the compiler would try to dereference an llvm::StructType.

Also slightly tweaked ther helper methods.