Allow no-prototype function's return type to a MMSafe pointer.
Checked C does not allow the return typf of a no-prototype function to
be a checked pointer type (by the time this rule was made Checkecd C
only have spatial-memory-safey pointers.  Now for the sake of easy
development, we allow MMSafe pointers to be returnd by a no-prototype
function.  Since the MMSafe pointer should also ensure spatial memory
safety (Future Work), we should fix this later.