Updated the dynamic ID checker to handle transparent struct object ID.
Previously we assume programmers manually add an ID field in a struct
in the source code; however, this is infeasible for array objects.
Now we make the ID of a heap object transparent to programmers:
when allocating a heap object, allocate 8 more bytes (for x86-64)
as the ID right before the object. Because the raw pointer still
points to the real start address of the heap object, during the dynamic
checking, we need first locate the ID by doing pointer arithemtic,
and then load the ID to compare it with the ID in the MMSafe pointer.

We use IRBuilder to construct new instructions; it is more convenient
than using CGBuilderTy as an IRBuilder usually requires fewer arguments
for instruction constructors.

The corresponding change in the customized memory allocator/deallocator
is at
https://github.com/jzhou76/checkedc-safe-mm-misc/commit/9b880101b68430cc1115d00741826602c57bd0c3