Implemented dynamic ID checking for _MM_array_ptr dereference.
Expanded the dynamic ID checking function for _MM_ptr dereference
to handle dynamic ID checking for _MM_array_ptr. Since there is a
dedicated pointer inside a _MM_array_ptr pointing to the ID of
an array, the dynamic checking procedure is simpler than the
dyanmic ID checking for _MM_ptr because there is no need to insert
extra instructions to locate the ID of a heap object. See the code
for more details.