Handle calling a function with generic _MM_array_ptr<T> parameters.
LLVM tends to flatten struct parameters to scalar types. When calling
a function with generic _MM_array_ptr<T> parameters, there would be
a type mismatch between the generic pointer type (i8*) and the concrete
pointer type. Here we mutate the concrete type to be the generic one.

The relavent commit for handling _MM_ptr is:
https://github.com/microsoft/checkedc-clang/pull/653/commits/61f3bf975daa1de6526166b0dce363f4152ebc7e