Changed dangerous mutateType() to type cast for generic function calls.
When calling a function with MMSafe pointer(s) to the generic type,
e.g., _MM_array_ptr<T>, as parameter or return value, there could be
a type mismatch between the concrete type and the generic type.
In the old implementation we used the llvm::Value::mutateType() method
to "solve" the type mismatch problem. However, this method is
very dangerous as it changes the type of a value forever and the
change take effects wherever the value is used. Quote from the
documentation of the llvm::Value class:

"Note that this is an extremely dangerous operation which can create
completely invalid IR very easily. It is strongly recommended that you
recreate IR objects with the right types instead of mutating them in
place."

In this new implementation we use cast instructions to create
new instructions to use. See the code for more details.

The old implementations are available:
for _MM_ptr<T> as return value:
https://github.com/microsoft/checkedc-clang/pull/653/commits/2209f31f7cd8fe5bc5304fbda653d391db40a51c
for _MM_ptr<T> as parameter:
https://github.com/microsoft/checkedc-clang/pull/653/commits/61f3bf975daa1de6526166b0dce363f4152ebc7e
for _MM_array_ptr<T> as return value:
https://github.com/jzhou76/checkedc-clang/commit/44d8e89659ad52c7ac9814f00ccf94acc3c9da69
for _MM_array_ptr<T> as parameter:
https://github.com/jzhou76/checkedc-clang/commit/39e4588716d5b72b19c0c3684218f3cb310e3a21