Comparing a MMSafePtr with MMSafePtr/NULL or cast NULL to an MMSafePtr.
When comparing an MMSafe pointer with NULL pointer or another
MMSafe pointer, we need extract the inner raw pointer to do the
comparison. Similarly when casting NULL to an MMSafe pointer, extract
the inner raw pointer for casting.

For example, "if (p == NULL) ..." where p is { %struct.node*, i64 }
would be compiled to (-O0):
  %_innerPtr = extractvalue { %struct.node*, i64 } %0, 0
  %cmp = icmp eq %struct.node* %_innerPtr, null

"if (p0 == p1) ..." would be compiled to:

   %_innerPtr = extractvalue { %struct.node*, i64 } %0, 0
   %_innerPtr1 = extractvalue { %struct.node*, i64 } %1, 0
   %cmp = icmp eq %struct.node* %_innerPtr, %_innerPtr1

To-Do: currently assigning a raw pointer to an MMSafe pointer to
the same type of struct causes compilation failure in the
IR code generation phase; however, this error should be caught earlier.