Improved code of handling casting to MMSafe pointer.
Casting between one type of MMSafe pointer, e.g., _MM_ptr<Node> to the
same type of MMSafe pointer to a different type, e.g., _MM_ptr<NewNode>,
can be dangerous. However, we should support this as sometimes the
programmer is certain that the cast is safe. For example, in the bh
program of the Olden benchmark suite
(https://github.com/zhaishuang1/test-suite/blob/master/MultiSource/Benchmarks/Olden/bh/defs.h)
it is safe to cast a pointer of "struct tree to a pointer of
struct node because struct node behaves like the parent "class" of
struct tree.

Currently we only allow casting a constant NULL pointer or another
MMSafe pointer to an MMSafe pointer. Casting other types to a MMSafe
pointer type should be caught as an error earlier during compiling.