Disallow cast a non-MMSafe pointer type to an MMSafe pointer type.
Casting a non-MMSafe pointer type to an MMSafe pointer type should be
forbidden. We let the compiler to catch this error during parsing.

On the other hand, we may need to support casting an MMSafe pointer
to a raw C pointer because sometimes we need pass an MMSafe pointer
to a unconverted original C function that accepts raw C pointer
arguments. That said, this is not the only way to achieve such
interoperability. Something like a bounds-safe inferface may serve
the same purpose without requiring programmers to write an explicit
cast from an MMSafe pointer to a raw one.