Implemented pointer arithmetic for MMArrayPtr (To-Do: ++/-- operators).
Implemented pointer arithemtic for MMArrayPtr. Essentially, the compiler
extracts the inner raw C pointer(s) and does a normal pointer
arithmetic. For pointer adding/subtracting a number, it creates
a new MMArrayPtr with the result from the normal pointer arithmetic,
and for subtraction between two MMArrayPtr, it returns the result
directly.

TO-DO:
1. ++/-- operators.
2. Disallow subtraction between two MMArrayPtr to different arrays.