Fixed a bug of checking an MMSafePtr inside a struct.
CodeGenFunction::EmitMemberExpr() calls EmitDynamicKeyCheck()
with the argument of the BaseExpr of the strut field.
When dereferencing an mm_ptr to a struct to access one of its fields,
such as "p->num", the BaseExpr represents "p" and EmitDynamicKeyCheck()
calls CodeGenFunction::EmitDeclRefLValue() to get an LValue of p.
However, when dereferencing an MMSafePtr inside a struct, e.g.,
"p->p0->num", EmitMemberExpr() passes the BaseExpr of num to
EmitDynamicKeyCheck(), which is p0. However, p0 itself is a MemberExpr,
and EmitDynamicKeyCheck() recursively calls EmitMemberExpr() to
get an LValue for p1. In EmitMemberExpr(), the old implementation
recursively calls EmitDynamicKeyCheck() with the BaseExpr of p1 - which
is p. This caused generating the check on p twice while missed checking
p1. This commit fixes this problem by adding one more parameter
to EmitMemberExpr() to indicate if it needs to call
EmitDynamicKeyCheck(). In summary, in EmitDynamicKeyCheck(), when
EmitMemberExpr() is called, set the flag to false to avoid
EmitMemberExpr() recursively calling EmitDynamicKeyCheck().