In parsing handle assigning a pointer created from '&' to an MMSafePtr.
We disallowed assigning a raw C pointer (except "NULL") to an MMSafe
pointer in commit https://github.com/jzhou76/checkedc-clang/commit/27cd060e69e00f377c07b766603e968fbd26e721.
However, this would break the compilation when the program creates a
raw pointer by using the address-of operator '&' to get the address
of an inner memory object inside a memory object pointed by an
MMSafe pointer.

This commit fixed this issue by checking if the memory object whose
address is taken is part of a memory object pointed by an MMSafe
pointer. Note this only fixed the issue at the parsing phase; we
still need to update the IR code generator to generate an MMSafe pointer
from the '&' operator.