Revised the code of assigning NULL to an MMSafePtr in EmitStoreOfScalar.
The old implementation (https://github.com/jzhou76/checkedc-clang/commit/88e4ee7b84b7389b866de4073729589e1a69dee0)
checks if the type of the RHS equals to the type of the LHS which
is an MMSafe pointer. When the compiler reaches this point, it means
the RHS is either a NULL or another MMSafePtr whose pointee's type
is the same as the type of the LHS's pointee because assigning a
non-MMSafePtr or MMSafePtr pointing to a different type would be
caught as compiling error during parsing. However, the code of the old
implementation is not self-explanatory. The new code explicitly
checks if the RHS represents a NULL and hence it is clearer to read.