Changed the dynamic key check by the 32-32 key-offset design for mm_ptr.
The old implementation for checking the validity of an mm_ptr assumes
that the second field of the pointer is totally for a 64-bit key.
However, we need support the situation that a '&' operator gets
the address of an inner field of a struct pointed to by an mm_ptr.
Our current implementation adopts a 32-32 key-offset design for the
metadata field for an mm_ptr. This commits changed the dynamic checking
for mm_ptr.

In the paper we say we support at least two ways of key-offset
allotment: 32-32 and 40-24. We may need to add support for the latter.

Also we may want to explore which is faster: key-offset or offset-key.