Fixed a bug in parsing an adddress-of operator expression.
If the access path in the expression of an '&' operator contains
a raw C pointer after an mmsafe pointer, e.g., p0 = &p->rp->obj where
p is an mmsafe pointer while rp is a raw pointer, and p0 is a mmsafe
pointer, it should be illegal because the address retrieved from the
expression is in an object pointed by a raw C pointer. The old
implementation allows this situation thus it is a bug.