Handled '&' operator on an object inside an object pointed by MMSafePtr.
If the '&' operator gets the address of an object inside an object
pointed by an MMSafe pointer, the result should be assigned to
an MMSafe pointer with the same key and the offset of the inner object.
In ScalarExprEmitter::VisitUnaryAddrOf(), we added code to handle
this situation.

TO-DO: The current implementation makes two assumptions specially for
an address-of expression that may not hold in certain programs.
First, it assumes there are only one or two indices in an GEP
instruction. Second, it assumes the first index of an GEP instruction
is zero. We have added assertions for this two situations.

TO-DO: the current code does not work for the current implementation
of mm_array_ptr. We plan to change the implementation of the
mm_array_ptr to the same as mm_ptr because it still gives us up
to 4GB offset and an 32-bit key, which should be sufficient for
most programs.  We can add one more pointer dedicated for enormous
arrays that are larger than 4GB.