Moved the dynamic key check instructions to a single function.
In the old implementation for each MMSafe pointer dereference the
compiler directly inserts a sequence of instructions to do the
key-lock checking. This commit puts all such instructions into a
single function and creates a call to this function.

For each compile module it generates one function for mm_ptr and one
for mm_array_ptr. By default it means the compiler would generate
these two functions for each module, and thus it would cause multiple
definition linking problem. We prefixed the names of these two functions
with the current module's name (the source file name). This solved
the multiple definition problem. However, there is still a minor
problem.  With -O2 or -O3 these two functions will get inlined; but
without LTO, the linker may not delete the function bodies.
As a result, there would be two dead functions left in each module.
This causes a little code size overhead, but it is almost negligible:
with -O3 there are only 11 x86 instructions in the function for mm_ptr
and 6 instructions for mm_array_ptr.