Fixed a bug in parsing address-of expressions.
Last commit https://github.com/jzhou76/checkedc-clang/commit/894abc52c86266069b1b36f6878e3dbf1ef26ab5
about parsing an address-of expression with a raw C pointer before
the first checked pointer in the access chain has a bug that caused
failed compilation of expressions like "&p->arr[3]".
The reason is that sometimes it's hard to differentiate
an array pointer from just a pointer: ArraySubscriptExpr::getBase()
returns an Expr of PointerType.  For example, for "&p->arr[3]" where
arr is of type Node, after running the getBase() method on the
array Expr, it returns an Expr of PointerType "Node *".
This commit added code to check if the PointerType of an Expr
comes from an array or just a pointer.