Implemented the address-of operator '&' on an item of an array.
Currently our design returns an mmptr for the address-of operator '&'
on an item of an array pointed by an mm_array_ptr. Semantically
such a expression means getting the address of a certain item
(or an field of an item of a struct array); it makes senses
to create a checked pointer to a singlon memory object. I've seen
code (thttpd) that applis the "->" operator to deference the result
pointer. However, since C does not differentiate a pointer to a singlton
object from a pointer to an array, a program might want to do pointer
arithmetic on the result pointer. For such cases, there are two
solutions. First, if the result is never dereferenced, the address-of
expression can be rewritten as a pointer arithmetic expression,
e.g., "&p[i]" -> "p + i". The result is natually an mm_array_ptr.
Second, we can offer an explicit cast intrinsic to cast an mm_ptr
to an mm_array_ptr.