Fixed a bug in generating a mmptr for '&' on a struct field.
We forgot to set the "isMMPtr" to true when creating the StructType
for the MMPtr to the struct field on which the '&' operator is applied.
One error caused by missing this setting is when using '&' to create
an mmptr and assign it to another mmptr to a different type. For
example, "p1 = &p->i" where p1 points to a long while i is of type int.
This would cause the assertion
"assert(0 && "Casting improper type to MMSafe pointer.");"
in ScalarExprEmitter::VisitCastExpr() to fail.

TODO: make the "isMMPtr" and "isMMArrayPtr" fields of llvm::StructType
private and add setters to update them.