Fixed a bug in parsing an address-of expr during ptr assignment checking
There was a bug in parsing an address-of expression during type checking
for pointer assignments. The old implementation doesn't differentiate
the two cases: getting the address of a pointer field of a struct,
and getting the address of something pointed by a pointer that is
a field of a struct. For example, "p1 = &p->i" where p1 is of type
mm_ptr<int> and p is a checked pointer to a struct that has an integer
field i, and "p1 = &p->pi->i" where p1 is of type mm_ptr<int>, p
is an mm_ptr to a struct that has a pointer field pi to integer type.
The former is allowed but the former should be forbidden as the
address-of expression is evaluated to a raw pointer. The old
implementation would throw a compile error for the first case,
and this commit fixed this.