[parsing] Fixed a bug in parsing assignment involving an address-of expr.
We disallow assigning an mmsafe pointer to a raw pointer. There was a
bug: when the RHS is an mmsafe pointer created from an address-of
expression and the LHS is a raw pointer, the compiler didn't catch it
during parsing but raised an assertion failure during the IR code
generation. This commit fixed this problem and had the compiler catch
the bug during parsing.

TODO: Currently the RHSType of the result of an addres-of expression is
always a raw pointer. For example, the type of "&p->i" is "int *" if i
is an integer, regardless of the pointer type of p. This would give
programmers a misleading error message. For example, "int *pi = &p->i;"
would raise an error like

    "initializing 'int *' with an expression of incompatible type 'int *'".

To fix this issue, we need set the type of the RHS correctly at a
much earlier stage.