Revised the address-of expression generation code.
During parsing an address-of expression, this commit moved the CastExpr
and ParenExpr removal loop from before to after the switch statement.
This would strip off ImplicitCastExpr of PointerType on ArrayType and
save the effort of checking if an PointerType expression is really a
pointer or an array. See the code for more details.