Updated the address-of expr generation code to handle ptr to union.
Added support to generate address-of expressions that involve mmsafe ptr
to a union. The change is not complex because a pointer accessing a
union member is essentially a bit cast from the pointer to the union
type to the member type. We just need to strip off certain bit casts
during the IR code generation for address-of expressions.

See the code for more details.