Fixed a bug in address-of expression generation.
While parsing an address-of expression, it should differentiate two
cases where the Expr's type is of mmsafe pointer: getting the address
of an mmsafe pointer and getting the address of something pointed by
an mmsafe pointer. For example, "mm_ptr<mm_ptr<int>> p = &p1;" where
p1 itself is an mm_ptr<int> should be disallowed. This can be achieved
by checking if the current Expr is the rightmost element in the access
chain. See the code for more details.