[WIP] Added the _multiple qualifier and generate code for '&' expr.
For local and global objects to which pointers may also point to the
heap, we added a new qualifer to them so that the compiler can generate
an mmsafe ptr to them when their addresses are taken. For example,
a function "void foo(int *p)" sometimes would like to take an
mm_ptr<int> to an integer on the heap, and sometimes it would like to
take a pointer derived from taking the address of an integer on the
stack. In this case, a programmer can add the new qualifer, temporarily
named "_multiple", to the declaration of the stack variable, and the
compiler will generate an mmptr from taking its address. For now we
assume that all _multple stack objects have a lock with value 1 and
all _multiple global objects have a lock with value 2.

TODO:
- Generate an mmptr for pointer arithmetic expressions of _multiple
  arrays.
- Generate initialization code to initialize the locks for
  _multiple objects.