[WIP][_multiple] Create a lock for _multiple stack variables.
At the end of generating a function, check if this function has
_multiple stack variables, and if it has, replace each such variable
with a struct that contains a lock and the original variable. Also
replace all uses of the AllocaInst for the original variable with
a GEP to the variable in the newly added struct. The lock for all
_multiple variables is 1.

On the llvm end, we extended AllocaInst with a field to indicate
if an AllocaInst is for an _multiple stack object. The commit is
at https://github.com/jzhou76/checkedc-llvm/commit/9a233656bd0749b2d5cd57f48780595f39d6953a