[_multiple] Create an mm_array_ptr from arith-expr of an _multiple array
Added code to create an mm_array_ptr from an arithmetic expression
of an _multiple array.

Also added code in SemaExpr.cpp to disallow assigning the result of
getting the address of an _multiple object to a raw C pointer. However,
it is not completely clear to me whether we should disallow this.
Allowing it would complicate both the meaning of programs and
the implementation of the compiler. For the semantics of programs,
allowing it means getting the address of an _multiple may generate
either a raw C pointer or an mmsafe ptr, depending on the LHS of
an assignment. In addtition, what should be generated if there is
no assignment? For example, using the result in a printf().
Regarding compiler implementation, currently an address-of expression
of an _multiple generates an mm_ptr. Allowing the situation
aforementioned would complicate the compiler implementation because
the function of address-of expression generation only takes the
expression as the single argument without knowing where the result
would be assigned to.