Implemented assigning a string constant to an mm_array_ptr.
Added support of directly assigning a string constant to an
mm_array_ptr, e.g., "mm_array_ptr<char> = "hello". String constants
are private globals; so we treat them as _multiple globals - a lock
is created for each string constant, and dereferences of mm_array_ptr
to string constants will be coupled with a dynamic key check.

TO-DO: disallow directly assigning a string constant to an mm_ptr.