[union] Fixed a bug in generating ptr from an addr-of expr with union.
There is a bug in the commit of handling an addr-of expr with a union
pointed by a pointer (https://github.com/jzhou76/checkedc-clang/commit/1646964cbf31b48531a21fa48ca980179c1da276).
If no mmsafe ptr is found in the expression, the returned result
is the bitcasted one but not the original one. This commit fixed this
problem by returning the original result in such cases.
See the code for more details.