[NULL] Fixed a bug in assigning NULL to an mmsafe pointer.
In the old implementation (https://github.com/jzhou76/checkedc-clang/blob/safe-mm-base/lib/CodeGen/CGExprScalar.cpp#L2531)
when assigning a NULL to an mmsafe pointer, the compiler extracts the
inner raw C pointer and assigns a NULL to it. There is no problem with
this operation for directly assigning a NULL to an mmsafe pointer
argument, return value, a local variable, or item of an aggregate
object. But there is bug in ternary expression generation, e.g.,

  p1 = num > 0 ? p : NULL;

where p1 and p are both checked pointers. The compiler creates a
PHI node for p and NULL but it triggerred an assertion failure of type
mismatch in PHINode::setIncomingValue() because the NULL's type
is *p_inner_ptr while p's type is an mmsafe pointer. This commit
generates a full NULL mmsafe pointer in replace of setting its
inner raw pointer to NULL. Currently all the small test cases
and Olden benchmarks can compile and run.

Related commits of setting mmsafe pointer to NULL:
https://github.com/jzhou76/checkedc-clang/commit/88e4ee7b84b7389b866de4073729589e1a69dee0
https://github.com/jzhou76/checkedc-clang/commit/2dd871247b186d622041a17c1c9f4f08840b63db
https://github.com/jzhou76/checkedc-clang/commit/5c3b786270490b714635f26e02bd6a100d6aca48
https://github.com/jzhou76/checkedc-clang/commit/1b2e3d24b80dc97c71db01000a7accc682b44cf4
https://github.com/jzhou76/checkedc-clang/commit/be66cd71bf1cc756b386874669fbd84a790d7afe
https://github.com/jzhou76/checkedc-clang/commit/96da6d0b7e59e1fa349da251a9a54424b8d34fbf
https://github.com/jzhou76/checkedc-clang/commit/2d2a8d985065d1e886353e81f6ff604deedb496d
https://github.com/jzhou76/checkedc-clang/commit/ee876f3783a467f89aab66d31880892e5245e763