[ternary] Added type checking for ternary expr with addr-of expr.
We did not handle assigning a ternary expression that has address-of
expression(s) to a pointer. For example, in this line of code

    int *pi = num > 0 ? &p->i : NULL;

p is an mm_ptr<int>, and this should cause the type checker to abort
because &p->i returns an mmsafe ptr. The compiler before this commit
does not handle this situation and it triggers an assertion failure
in IR code generation.

Similar, for this line of code,

    mm_ptr<int> pi = num > 0 ? &p->i : NULL;

it should pass the type checking but the compiler threw a type mismatch
error because it thought the RHS returned a raw pointer.

This commit fixed these problems by handling ternary expressions with
address-of expressions as result(s). It parses both the true result
and the false result of an ternary expression.

See the code for more details.

TODO: Although the current compiler can handle the two examples
correctly during type checking, it throws an assertion failure
during IR code generation for the second example.