[ternary] Fixed a bug in generating a ternary expression involing NULL.
When generating a ternary expression and assigning the result to
an mmsafe pointer, if one side of the ternary expr is an addr-of expr
that returns an mmsafe ptr, and the other side is a NULL, then there
would a type mismatch between the two sides. For example, for

    mm_ptr<int> pi = num > 0 ? &p->i : NULL;

although the compiler generated the llvm type for (&p->i) correctly,
it has a problem that was acknowledged earlier but not fixed (https://github.com/jzhou76/checkedc-clang/commit/c1fae9585aa4308972568e5499bc9be71597d7e7#diff-b0c2a08569a0c3beb9cc403c2486a866570e86f5aac8092d39a25a270fa90e66R8370).
In short, if p is an mmsafe pointer then the result of &p->i
is an mmsafe pointer essentially but on the clang side it is still
maintained as a raw pointer. For the example above, this caused the
type of the NULL to be a "int *" instead of a NULL mmsafe pointer.
This commit fiexd this problem by generating a NULL mmsafe pointer
for the NULL.